+++ packages/core/lib/combo/settings/random.ts
@@ -254,12 +254,24 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
     base.soulsEnemyMm = true;
     base.soulsBossOot = true;
     base.soulsBossMm = true;
+    base.soulsNpcOot = true;
+    base.soulsNpcMm = true;
+    //base.soulsAnimalOot = true;
+    //base.soulsAnimalMm = true;
+    base.soulsMiscOot = true;
+    base.soulsMiscMm = true;
     break;
   default:
-    base.soulsEnemyOot = booleanWeighted(random, 0.5);;
-    base.soulsEnemyMm = booleanWeighted(random, 0.5);;
-    base.soulsBossOot = booleanWeighted(random, 0.5);;
-    base.soulsBossMm = booleanWeighted(random, 0.5);;
+    base.soulsEnemyOot = booleanWeighted(random, 0.5);
+    base.soulsEnemyMm = booleanWeighted(random, 0.5);
+    base.soulsBossOot = booleanWeighted(random, 0.5);
+    base.soulsBossMm = booleanWeighted(random, 0.5);
+    base.soulsNpcOot = booleanWeighted(random, 0.5);
+    base.soulsNpcMm = booleanWeighted(random, 0.5);
+    //base.soulsAnimalOot = booleanWeighted(random, 0.5);
+    //base.soulsAnimalMm = booleanWeighted(random, 0.5);
+    base.soulsMiscOot = booleanWeighted(random, 0.5);
+    base.soulsMiscMm = booleanWeighted(random, 0.5);
   }
 
   /* Events - 25% full open, 25% full closed, 50% individual */